ssize_t型の導入: 文字列長やファイルサイズの計算においてint型からssize_t型への変更を行い、特に大容量ファイルの処理における精度を向上させる修正を実施。これにより、32ビットおよび64ビット環境での互換性を確保。

+++ sakura_core/agent/CGrepAgent.cpp
@@ -263,7 +263,7 @@ void CGrepAgent::AddTail( CEditView* pcEditView, const CNativeW& cmem, bool bAdd
 					pcEditView->GetDocument()->GetDocumentEncoding(), 0) );
 			pcCodeBase->UnicodeToCode( cmem, &cmemOut );
 			DWORD dwWrite = 0;
-			::WriteFile(out, cmemOut.GetRawPtr(), cmemOut.GetRawLength(), &dwWrite, nullptr);
+			::WriteFile(out, cmemOut.GetRawPtr(), static_cast<DWORD>(cmemOut.GetRawLength()), &dwWrite, nullptr);
 		}
 	}else{
 		pcEditView->GetCommander().Command_ADDTAIL( cmem.GetStringPtr(), cmem.GetStringLength() );
