+++ sakura_core/charset/CEuc.h
@@ -32,9 +32,9 @@ public:
 	// 2008.11.10 変換ロジックを書き直す
 	inline static int _EucjpToUni_char( const unsigned char* pSrc, unsigned short* pDst, const ECharSet eCharset, bool* pbError, bool* pbHex );
 protected:
-	static int EucjpToUni( const char* pSrc, const int nSrcLen, wchar_t* pDst, bool* pbError );
+	static int EucjpToUni( const char* pSrc, const ssize_t nSrcLen, wchar_t* pDst, bool* pbError );
 	inline static int _UniToEucjp_char( const unsigned short* pSrc, unsigned char* pDst, const ECharSet eCharset, bool* pbError );
-	static int UniToEucjp( const wchar_t* pSrc, const int nSrcLen, char* pDst, bool* pbError );
+	static int UniToEucjp( const wchar_t* pSrc, const ssize_t nSrcLen, char* pDst, bool* pbError );
 };
 
 /*!
@@ -46,7 +46,7 @@ protected:
 */
 inline int CEuc::_EucjpToUni_char( const unsigned char* pSrc, unsigned short* pDst, const ECharSet eCharset, bool* pbError, bool* pbHex = nullptr )
 {
-	int nret;
+	ssize_t nret;
 	unsigned char czenkaku[2];
 	unsigned int ctemp;
 	bool berror=false;
@@ -101,7 +101,7 @@ inline int CEuc::_EucjpToUni_char( const unsigned char* pSrc, unsigned short* pD
 		*pbHex = hex;
 	}
 
-	return nret;
+	return static_cast<int>(nret);
 }
 
 /*
