+++ sakura_core/charset/CUtf8.h
@@ -42,10 +42,10 @@ protected:
 
 	//変換の実装
 	// 2008.11.10 変換ロジックを書き直す
-	inline static int _Utf8ToUni_char( const unsigned char* pSrc, const int nSrcLen, unsigned short* pDst, bool bCESUMode );
-	static int Utf8ToUni( const char* pSrc, const int nSrcLen, wchar_t* pDst, bool bCESU8Mode );
-	inline static int _UniToUtf8_char( const unsigned short* pSrc, const int nSrcLen, unsigned char* pDst, bool bCESU8Mode );
-	static int UniToUtf8( const wchar_t* pSrc, const int nSrcLen, char* pDst, bool* pbError, bool bCESU8Mode );
+	inline static ssize_t _Utf8ToUni_char( const unsigned char* pSrc, const ssize_t nSrcLen, unsigned short* pDst, bool bCESUMode );
+	static ssize_t Utf8ToUni( const char* pSrc, const ssize_t nSrcLen, wchar_t* pDst, bool bCESU8Mode );
+	inline static int _UniToUtf8_char( const unsigned short* pSrc, const ssize_t nSrcLen, unsigned char* pDst, bool bCESU8Mode );
+	static int UniToUtf8( const wchar_t* pSrc, const ssize_t nSrcLen, char* pDst, bool* pbError, bool bCESU8Mode );
 };
 
 /*!
@@ -56,9 +56,9 @@ protected:
 	高速化のため、インライン化
 
 */
-inline int CUtf8::_Utf8ToUni_char( const unsigned char* pSrc, const int nSrcLen, unsigned short* pDst, bool bCESUMode )
+inline ssize_t CUtf8::_Utf8ToUni_char( const unsigned char* pSrc, const ssize_t nSrcLen, unsigned short* pDst, bool bCESUMode )
 {
-	int nret;
+	ssize_t nret;
 
 	if( nSrcLen < 1 ){
 		return 0;
@@ -106,7 +106,7 @@ inline int CUtf8::_Utf8ToUni_char( const unsigned char* pSrc, const int nSrcLen,
 
 	高速化のため、インライン化
 */
-inline int CUtf8::_UniToUtf8_char( const unsigned short* pSrc, const int nSrcLen, unsigned char* pDst, bool bCESU8Mode )
+inline int CUtf8::_UniToUtf8_char( const unsigned short* pSrc, const ssize_t nSrcLen, unsigned char* pDst, bool bCESU8Mode )
 {
 	int nret;
